fix: achieve zero-warning build with strict error checking
- Enable -Werror to treat warnings as errors
- Fix const-correctness and sign comparison warnings
- Add proper braces for VkClearColorValue initializers
- Apply relaxed warnings to external libraries only
- Maintain strict standards for project source code

All builds now complete with zero warnings and zero errors.

+++ src/vk/vk_physical_device.c
@@ -34,7 +34,7 @@ static void deletePhysicalDevices(VkPhysicalDevice *physicalDevices){
 
 static uint32_t getPhysicalDeviceTotalMemory(const VkPhysicalDeviceMemoryProperties *pPhysicalDeviceMemoryProperties){
 	uint32_t physicalDeviceTotalMemory = 0;
-	for(int i = 0; i < pPhysicalDeviceMemoryProperties->memoryHeapCount; i++){
+	for(int i = 0; i < (int)pPhysicalDeviceMemoryProperties->memoryHeapCount; i++){
 		if((pPhysicalDeviceMemoryProperties->memoryHeaps[i].flags & VK_MEMORY_HEAP_DEVICE_LOCAL_BIT) != 0){
 			physicalDeviceTotalMemory += pPhysicalDeviceMemoryProperties->memoryHeaps[i].size;
 		}
